requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
Brotli==1.1.0
certifi==2024.2.2
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
pillow==10.3.0
sydney-py==0.20.4
websockets==12.0
yarl==1.9.4
